+++ commands/avatar.js
@@ -4,7 +4,8 @@ module.exports = {
     args: false,
     usage: '',
     guildOnly: false,
-    adminOnly: false,
+    moderatorOnly: false,
+    creatorOnly: false,
     cooldown: 0,
     execute(message, args) {
         if (!message.mentions.users.size) {
