+++ commands/listening.js
@@ -4,7 +4,8 @@ module.exports = {
     args: false,
     usage: '',
     guildOnly: false,
-    adminOnly: true,
+    moderatorOnly: false,
+    creatorOnly: true,
     cooldown: 60,
     execute(message, args) {
         process.env.BOT_TOKEN ?
