+++ commands/quote.js
@@ -7,7 +7,8 @@ module.exports = {
     description: 'Affiche une citation du stream de Bob choisie au hasard.',
     args: false,
     guildOnly: false,
-    adminOnly: false,
+    moderatorOnly: false,
+    creatorOnly: false,
     cooldown: 0,
     execute(message, args) {
             const url = `https://twitch.center/customapi/quote/list?token=ecd68f5b`;
