+++ commands/test.js
@@ -6,7 +6,8 @@ module.exports = {
     args: false,
     usage: '',
     guildOnly: true,
-    adminOnly: true,
+    moderatorOnly: false,
+    creatorOnly: true,
     cooldown: 60,
     execute(message, args) {
