+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "botzilla",
-  "version": "0.9.0",
+  "version": "1.0.0",
   "description": "Discord bot for Bobzilla4tv server",
   "main": "bot.js",
   "scripts": {
